requirements.txt
numpy==1.19.2
Pillow==9.1.1
streamlit==1.10.0
tensorflow==1.14
opencv-python-headless
